v1.1.0
Améliorations/Corrections du code et ajout d'un rapport mensuel avec des statistiques

+++ meteocheck.py
@@ -2,7 +2,7 @@
 *
 * PROJET : MeteoCheck
 * AUTEUR : Arnaud R.
-* VERSIONS : 1.0.0
+* VERSIONS : 1.1.0
 * NOTES : None
 *
 '''
@@ -53,9 +53,9 @@ sent_alerts = {
     'pressure': None
 }
 
-async def log_uncaught_exceptions(exc_type, exc_value, exc_traceback):
+def log_uncaught_exceptions(exc_type, exc_value, exc_traceback):
     error_message = "".join(traceback.format_exception(exc_type, exc_value, exc_traceback))
-    await log_message(f"Uncaught exception: {error_message}")
+    asyncio.run(log_message(f"Uncaught exception: {error_message}"))
 
 sys.excepthook = log_uncaught_exceptions
 
@@ -65,10 +65,13 @@ async def log_message(message: str):
 
 async def get_weather_data():
     try:
+        # Arrondir l'heure actuelle à l'heure précise dès le début
+        now = pd.Timestamp.now(tz='Europe/Berlin').floor('H')
+
         async with aiohttp.ClientSession() as session:
             async with session.get(weather_url) as resp:
                 data = await resp.json()
-    
+
                 # Extraire les données nécessaires
                 time = data['hourly']['time']
                 temperature = data['hourly']['temperature_2m']
@@ -88,39 +91,48 @@ async def get_weather_data():
                     'uv_index': uv_index,
                     'pressure_msl': pressure_msl   
                 })
-                
+
                 # Convertir le temps unix en un objet datetime pour faciliter les comparaisons
                 df['time'] = pd.to_datetime(df['time'], unit='s').dt.tz_localize('Europe/Berlin')
                 # Arrondir les heures dans la DataFrame
                 df['time'] = df['time'].dt.floor('H')
 
-                # Arrondir l'heure actuelle à l'heure précise et sélectionner les données correspondantes
-                now = pd.Timestamp.now(tz='Europe/Berlin').floor('H')
+                # Lire le fichier csv existant
+                df_existing = pd.read_csv(csv_filename)
+                df_existing['time'] = pd.to_datetime(df_existing['time'])
+
                 # Filtrer les données pour garder seulement celles de l'heure précédente
                 hour_ago = now - pd.Timedelta(hours=1)
                 past_hour_df = df[(df['time'] >= hour_ago) & (df['time'] < now)]
-                # Filtrer les données pour garder seulement des 7 prochaines heures
-                six_hours_later = now + pd.Timedelta(hours=7)
-                next_six_hours_df = df[(df['time'] >= now) & (df['time'] < six_hours_later)]
-
-                try:
-                    # Vérifiez si le fichier existe pour déterminer si nous devons écrire l'en-tête
-                    write_header = not os.path.exists(csv_filename)
-
-                    # Enregistrer les données dans un CSV
-                    past_hour_df.to_csv(csv_filename, mode='a', header=write_header, index=False)
                 
-                except Exception as e:
-                    await log_message(f"Error in writing data to csv in get_weather_data: {str(e)}")           
+                # Vérifier si l'heure précédente a déjà été enregistrée
+                past_hour_exists_in_csv = not df_existing[df_existing['time'] == hour_ago].empty
+                if not past_hour_exists_in_csv:
+                    # Si l'heure précédente n'a pas encore été enregistrée, ajouter seulement les données de l'heure précédente
+                    try:
+                        # Vérifiez si le fichier existe pour déterminer si nous devons écrire l'en-tête
+                        write_header = not os.path.exists(csv_filename)
+
+                        # Enregistrer les données dans un CSV
+                        past_hour_df.to_csv(csv_filename, mode='a', header=write_header, index=False)
+                
+                    except Exception as e:
+                        await log_message(f"Error in writing data to csv in get_weather_data: {str(e)}")
+
+                # Filtrer les données pour garder seulement des 7 prochaines heures
+                seven_hours_later = now + pd.Timedelta(hours=7)
+                next_seven_hours_df = df[(df['time'] > now) & (df['time'] < seven_hours_later)]
 
                 # Retourner les données pour les utiliser dans d'autres fonctions
-                return next_six_hours_df
-    
+                return next_seven_hours_df
+
     except Exception as e:
         await log_message(f"Error in get_weather_data: {str(e)}")
         return pd.DataFrame()
     
-async def send_alert(message):
+async def send_alert(message, row=None, alert_column=None):
+    if row is not None and alert_column is not None:
+        await check_records(row, alert_column)
     try:
         # Lire le fichier json pour obtenir tous les IDs de chat
         with open('chat_ids.json', 'r') as file:
@@ -153,25 +165,25 @@ async def check_weather():
             if row['temperature'] > 35 or row['temperature'] < -10:
                 # Vérifier si une alerte a déjà été envoyée pour ce jour
                 if sent_alerts['temperature'] != time.date():
-                    await send_alert(f"Alerte météo : Température prévue de {row['temperature']}°C à {time}.")
+                    await send_alert(f"Alerte météo : Température prévue de {row['temperature']}°C à {time}.", row, 'temperature')
                     sent_alerts['temperature'] = time.date()
             
             # Vérifier les conditions de précipitations
             if row['precipitation_probability'] > 80 and row['precipitation'] > 5:
                 if sent_alerts['precipitation'] != time.date():
-                    await send_alert(f"Alerte météo : Précipitations prévues de {row['precipitation']}mm à {time}.")
+                    await send_alert(f"Alerte météo : Précipitations prévues de {row['precipitation']}mm à {time}.", row, 'precipitation')
                     sent_alerts['precipitation'] = time.date()
 
             # Vérifier les conditions de vent
             if row['windspeed'] > 30:
                 if sent_alerts['windspeed'] != time.date():
-                    await send_alert(f"Alerte météo : Vent prévu de {row['windspeed']}km/h à {time}.")
+                    await send_alert(f"Alerte météo : Vent prévu de {row['windspeed']}km/h à {time}.", row, 'windspeed')
                     sent_alerts['windspeed'] = time.date()
 
             # Vérifier les conditions d'index UV
             if row['uv_index'] > 8:
                 if sent_alerts['uv_index'] != time.date():
-                    await send_alert(f"Alerte météo : Index UV prévu de {row['uv_index']} à {time}.")
+                    await send_alert(f"Alerte météo : Index UV prévu de {row['uv_index']} à {time}.", row, 'uv_index')
                     sent_alerts['uv_index'] = time.date()
 
             # Vérifier les conditions de pression
@@ -179,39 +191,94 @@ async def check_weather():
                 pressure_drop = df['pressure_msl'].iloc[0] - df['pressure_msl'].iloc[5]
                 if pressure_drop >= 1.2:
                     if sent_alerts['pressure'] != time.date():
-                        await send_alert(f"Alerte météo : Baisse rapide de la pression atmosphérique prévue de {pressure_drop} hPa sur 7 heures à partir de {time}.")
+                        await send_alert(f"Alerte météo : Baisse rapide de la pression atmosphérique prévue de {pressure_drop} hPa sur 7 heures à partir de {time}.", row, 'pressure')
                         sent_alerts['pressure'] = time.date()
     
     except Exception as e:
         await log_message(f"Error in check_weather: {str(e)}")    
 
-async def end_of_year_summary():
+async def check_records(row, alert_column):
+    # Lire toutes les données
+    df = pd.read_csv(csv_filename)
+
+    # si le df est vide, on peut simplement retourner 
+    if df.empty:
+        await log_message(f"No data found in the csv file.")
+        return
+
+    # Convertir 'time' en datetime
+    df['time'] = pd.to_datetime(df['time'])
+
+    # Obtenir l'année de la ligne actuelle
+    current_year = row['time'].year
+
+    # Filtrer le dataframe pour n'inclure que l'année actuelle
+    df = df[df['time'].dt.year == current_year]
+
+    # Vérifier le record pour la colonne d'alerte
+    max_value = df[alert_column].max()
+    if row[alert_column] > max_value:
+        await send_alert(f"Alerte météo : Nouveau record de {alert_column} à {row[alert_column]} à {row['time']}.")
+
+async def end_of_month_summary():
     try:
-        # Lire toutes les données
         df = pd.read_csv(csv_filename)
-
-        # si le df est vide, on peut simplement retourner 
-        if df.empty:
-            await log_message(f"Error in end_of_year_summary: No data found in the csv file.")
-            return
-
-        # Convertir 'time' en datetime
         df['time'] = pd.to_datetime(df['time'])
 
-        # Filtrer pour la dernière année
-        df = df[df['time'].dt.year == df['time'].dt.max().year]  # type: ignore
+        current_month = df['time'].dt.to_period('M').max()
+        df = df[df['time'].dt.to_period('M') == current_month]
 
+        df.set_index(df['time'].dt.date, inplace=True)
 
-        # Trouver les valeurs max et min
-        max_values = df.max()
-        min_values = df.min()
+        hot_day = df['temperature'].idxmax()
+        max_temp = df.loc[hot_day, 'temperature'].max() # type: ignore
 
-        # Supprimer les valeurs indésirées
-        max_values = max_values.drop(['precipitation_probability', 'precipitation'])
-        min_values = min_values.drop(['precipitation_probability', 'precipitation', 'windspeed'])
+        cold_day = df['temperature'].idxmin()
+        min_temp = df.loc[cold_day, 'temperature'].min() # type: ignore
 
-        # Envoyer le résumé
-        await send_alert(f"Résumé de fin d'année : Valeurs maximales - {max_values}. Valeurs minimales - {min_values}.")
+        rain_day = df['precipitation'].idxmax()
+        max_precipitation = df.loc[rain_day, 'precipitation'].max() # type: ignore
+
+        uv_day = df['uv_index'].idxmax()
+        max_uv_index = df.loc[uv_day, 'uv_index'].max() # type: ignore
+
+        rain_threshold = 0.1
+        rainy_days = df[df['precipitation'] > rain_threshold].index.nunique()
+
+        await send_alert(f"Résumé de fin de mois : Le jour le plus chaud était {hot_day} avec {max_temp}°C. Le jour le plus froid était {cold_day} avec {min_temp}°C. Le jour avec le plus de pluie était {rain_day} avec {max_precipitation}mm. Le jour avec le plus fort index UV était {uv_day} avec un index de {max_uv_index}. Le nombre de jours de pluie était {rainy_days}.")
+
+    except Exception as e:
+        await log_message(f"Error in end_of_month_summary: {str(e)}")
+
+async def end_of_year_summary():
+    try:
+        df = pd.read_csv(csv_filename)
+        df['time'] = pd.to_datetime(df['time'])
+
+        current_year = df['time'].dt.to_period('Y').max()
+        df = df[df['time'].dt.to_period('Y') == current_year]
+
+        df.set_index(df['time'].dt.date, inplace=True)
+
+        hot_day = df['temperature'].idxmax()
+        max_temp = df.loc[hot_day, 'temperature'].max() # type: ignore
+
+        cold_day = df['temperature'].idxmin()
+        min_temp = df.loc[cold_day, 'temperature'].min() # type: ignore
+
+        rain_day = df['precipitation'].idxmax()
+        max_precipitation = df.loc[rain_day, 'precipitation'].max() # type: ignore
+
+        uv_day = df['uv_index'].idxmax()
+        max_uv_index = df.loc[uv_day, 'uv_index'].max() # type: ignore
+
+        wind_day = df['wind_speed'].idxmax()
+        max_wind_speed = df.loc[wind_day, 'wind_speed'].max() # type: ignore
+
+        rain_threshold = 0.1
+        rainy_days = df[df['precipitation'] > rain_threshold].index.nunique()
+
+        await send_alert(f"Résumé de fin d'année : Le jour le plus chaud était {hot_day} avec {max_temp}°C. Le jour le plus froid était {cold_day} avec {min_temp}°C. Le jour avec le plus de pluie était {rain_day} avec {max_precipitation}mm. Le jour avec le plus fort index UV était {uv_day} avec un index de {max_uv_index}. Le jour avec le vent le plus fort était {wind_day} avec une vitesse de {max_wind_speed} km/h. Le nombre de jours de pluie était {rainy_days}.")
 
     except Exception as e:
         await log_message(f"Error in end_of_year_summary: {str(e)}")
@@ -221,7 +288,9 @@ async def schedule_jobs():
         try:
             await check_weather()
             now = datetime.datetime.now()
-            if now.month == 12 and now.day == 31 and now.hour == 0:
+            if now.day == 1 and now.hour == 0:
+                await end_of_month_summary()
+            elif now.month == 12 and now.day == 31 and now.hour == 0:
                 await end_of_year_summary()
             await asyncio.sleep(3600)
         except Exception as e:
